Add files via upload
This is an example code for arduino based cold brew coffee make. 
It works with a solenoid valve and 2 rotary encoders to control water flow frequency and 2 switches to initiate and stopping brewing process which will turn on status leds.
The amount of brewed coffee is measured by hx711 weight sensor and the end of brewing is notified by a buzzer.